rrr: convert misc/serializable_envelope to module (rrr.serializable_envelope)
serializable_envelope.cpp module; .hpp shim. Exports
SerializableEnvelope<TypeList> template (closed-set polymorphic
carrier), pack_aliased / pack / unpack / unpack_shared methods,
and Marshal stream operators. Imports rrr.marshal for
MarshalSink/Source + make_*_proxy helpers; rrr.serializable for
SerializableProxy + the TypeList machinery.

Switched the Marshal stream operators to use make_*_proxy explicitly
(previously called the removed BinaryWriteArchive(MarshalSink*)
convenience constructor).

34 modules. 78.09 s wallclock = 4.63× baseline.
